upload-builder

+++ Builder.sh
@@ -20,13 +20,19 @@ if [ $option = 1 ]
         then 
         cd ./Sources/M1721/
         bash build-kernel.sh | tee /root/K-Nel-Projects/OutFiles/Build-Kernel-M1721.log
-        echo 'Pack Starting······'
         if [ -e '/root/K-Nel-Projects/OutFiles/K-Nel-M1721.zip' ]
           then rm '/root/K-Nel-Projects/OutFiles/K-Nel-M1721.zip'
         fi
-        cp -r out/arch/arm64/boot/Image.gz /root/K-Nel-Projects/Toolchain/Anykernel3/
-        cd /root/K-Nel-Projects/Toolchain/Anykernel3/
-        zip -r '/root/K-Nel-Projects/OutFiles/K-Nel-M1721.zip' *
+        if [ -e out/arch/arm64/boot/Inage.gz ]
+        then
+            echo 'Pack Starting······'
+            cp -r out/arch/arm64/boot/Image.gz /root/K-Nel-Projects/Toolchain/Anykernel3/
+            cd /root/K-Nel-Projects/Toolchain/Anykernel3/
+            zip -r '/root/K-Nel-Projects/OutFiles/K-Nel-M1721.zip' *
+        else
+        echo 'Build Failed!'
+        exit
+        fi
         if [ -s '/root/K-Nel-Projects/OutFiles/K-Nel-M1721.zip' ]
           then echo 'Pack Successed!'
           else
@@ -40,7 +46,7 @@ if [ $option = 1 ]
 fi
 
 if [ $option = 2 ]
-then bash ./Sources/upload-sources.sh
+then bash ./Sources/upload-sources.sh | tee ./OutFiles/Upload-Sources.log
 fi
 
 if [ $option = 3 ]
